Further Accuracy Improvement
This is almost as accurate as it needs to be. The last thing needed to improve the accuracy is to replace the log2() functions with functions that take the logarithm of a function with a base of 1.5 instead of 2.

The speed which produces the greatest distance still needs to be coded to print to the screen.